+++ src/index.js
@@ -3,12 +3,14 @@ export {default as TxSignature} from './tx-signature.js';
 export {default as TxMultisignature} from './tx-multisignature.js';
 export {default as TxData} from './tx-data/index.js';
 
+// coin operation
 export {default as TxDataSend} from './tx-data/send.js';
 export {default as TxDataMultisend} from './tx-data/multisend.js';
 export {default as TxDataSell} from './tx-data/sell.js';
 export {default as TxDataBuy} from './tx-data/buy.js';
 export {default as TxDataSellAll} from './tx-data/sell-all.js';
 
+// coin administration
 export {default as TxDataCreateCoin} from './tx-data/create-coin.js';
 export {default as TxDataRecreateCoin} from './tx-data/recreate-coin.js';
 export {default as TxDataEditTickerOwner} from './tx-data/edit-ticker-owner.js';
@@ -17,6 +19,7 @@ export {default as TxDataRecreateToken} from './tx-data/recreate-token.js';
 export {default as TxDataMintToken} from './tx-data/mint-token.js';
 export {default as TxDataBurnToken} from './tx-data/burn-token.js';
 
+// validator administration
 export {default as TxDataDeclareCandidacy} from './tx-data/declare-candidacy.js';
 export {default as TxDataEditCandidate} from './tx-data/edit-candidate.js';
 export {default as TxDataEditCandidatePublicKey} from './tx-data/edit-candidate-public-key.js';
@@ -24,19 +27,25 @@ export {default as TxDataEditCandidateCommission} from './tx-data/edit-candidate
 export {default as TxDataSetCandidateOn} from './tx-data/set-candidate-on.js';
 export {default as TxDataSetCandidateOff} from './tx-data/set-candidate-off.js';
 
+// stake operation
 export {default as TxDataDelegate} from './tx-data/delegate.js';
 export {default as TxDataUnbond} from './tx-data/unbond.js';
 export {default as TxDataMoveStake} from './tx-data/move-stake.js';
+export {default as TxDataLockStake} from './tx-data/lock-stake.js';
 
+// misc
 export {default as TxDataRedeemCheck} from './tx-data/redeem-check.js';
 export {default as TxDataCreateMultisig} from './tx-data/create-multisig.js';
 export {default as TxDataEditMultisig} from './tx-data/edit-multisig.js';
+export {default as TxDataLock} from './tx-data/lock.js';
 
+// dao
 export {default as TxDataSetHaltBlock} from './tx-data/vote-halt-block.js';
 export {default as TxDataPriceVote} from './tx-data/vote-price.js';
 export {default as TxDataVoteCommission} from './tx-data/vote-commission.js';
 export {default as TxDataVoteUpdate} from './tx-data/vote-update.js';
 
+// dex
 export {default as TxDataCreateSwapPool} from './tx-data/create-swap-pool.js';
 export {default as TxDataAddLiquidity} from './tx-data/add-liquidity.js';
 export {default as TxDataRemoveLiquidity} from './tx-data/remove-liquidity.js';
@@ -44,8 +53,10 @@ export {default as TxDataBuySwapPool} from './tx-data/buy-swap-pool.js';
 export {default as TxDataSellSwapPool} from './tx-data/sell-swap-pool.js';
 export {default as TxDataSellAllSwapPool} from './tx-data/sell-all-swap-pool.js';
 
+// limit order
 export {default as TxDataAddLimitOrder} from './tx-data/add-limit-order.js';
 export {default as TxDataRemoveLimitOrder} from './tx-data/remove-limit-order.js';
+
 export {coinToBuffer, bufferToCoin} from './helpers.js';
 export {default as defineProperties} from './define-properties.js';
 export {TX_TYPE, txTypeList, normalizeTxType} from './tx-types.js';