+++ test/tx-data/lock-stake.test.js
@@ -1,0 +1,20 @@
+import {TX_TYPE, toBuffer} from 'minterjs-util';
+import {TxData, TxDataLockStake} from '~/src';
+import decodeToArray from '../decode-to-array';
+
+describe('TxDataLockStake', () => {
+    const params = {
+    };
+
+    const serializedTxData = (new TxDataLockStake(params)).serialize();
+
+    test('rlp encoded fields', () => {
+        expect(decodeToArray(serializedTxData))
+            .toEqual([
+            ]);
+    });
+
+    test('TxData', () => {
+        expect(TxData(params, TX_TYPE.LOCK_STAKE).serialize()).toEqual(serializedTxData);
+    });
+});