+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## WIP
+- add `Lock` and `LockStake` tx data
+- **BREAKING** add fields to VoteCommission tx data
+
 ## 11.1.0 - 2021.12.24
 - Tx: add `getSenderAddressString` method